requirements.txt
fastapi
uvicorn[standard]
celery[redis,sqlite]
sqlalchemy
python-jose[cryptography]
passlib[bcrypt]
python-dotenv